refactor(ui): share session-default reading
