+++ ui/src/lib/sessions/navigation.ts
@@ -15,6 +15,7 @@ import {
   normalizeAgentId,
   normalizeSessionKeyForUiComparison,
   parseAgentSessionKey,
+  readSessionDefaults,
   resolveUiConfiguredMainKey,
   resolveUiDefaultAgentId,
   resolveUiGlobalAliasAgentId,
@@ -62,23 +63,6 @@ export type SessionScopeHostWithKey = SessionScopeHost & {
 };
 
 export type SessionRefreshTarget = { sessionKey: string; agentId?: string };
-
-type SessionDefaults = {
-  defaultAgentId?: string | null;
-  mainKey?: string | null;
-  mainSessionKey?: string | null;
-};
-
-function readSessionDefaults(
-  host: Pick<SessionNavigationInput, "hello">,
-): SessionDefaults | undefined {
-  const snapshot = host.hello?.snapshot;
-  if (!snapshot || typeof snapshot !== "object" || !("sessionDefaults" in snapshot)) {
-    return undefined;
-  }
-  const defaults = snapshot.sessionDefaults;
-  return defaults && typeof defaults === "object" ? (defaults as SessionDefaults) : undefined;
-}
 
 export function resolveSessionKey(
   sessionKey: string | undefined | null,
